requirements.txt
app==0.0.1
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.9
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.3
joblib==0.16.0
kiwisolver==1.2.0
lxml==4.6.3
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.5
parsel==1.6.0
Pillow==8.2.0
pymongo==3.10.1
pyparsing==2.4.7
python-application==2.8.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4
regex==2020.7.14
requests==2.23.0
requests-file==1.5.1
scipy==1.5.1
seaborn==0.10.1
selectorlib==0.16.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.48.2
urllib3==1.26.5
w3lib==1.22.0
Werkzeug==1.0.1
zope.interface==5.1.0
gunicorn==20.0.4
